Improve safety of delete account endpoint

+++ app/deleteAccount.py
@@ -4,10 +4,15 @@ from .models import User
 
 def deleteAccount():
     username = request.json["username"]
+    password = request.json["password"]
+
     user = User.query.filter_by(username = username).first()
 
     if user is None:
         return jsonify({"error": "User not found"}), 404
+
+    if not user.check_password(password):
+        return jsonify({"error": "Incorrect password"}), 400
 
     # Cascades to sensors and measurements
     db.session.delete(user)
